# public/swagger.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "ct-project-api",
        "version": "0.1"
    },
    "servers": [
        {
            "url": "https://ct-project.pp.ua/api/"
        },
        {
            "url": "http://172.19.0.1/api/"
        }
    ],
    "paths": {
...
        "/v1/admin/cities": {
            "get": {
                "tags": [
                    "Admin Panel"
                ],
                "operationId": "c91f673c41f31d221cd1ab64f2fdc0a7",
                "parameters": [
                    {
                        "name": "lastId",
                        "in": "query",
                        "description": "Min: 0",
                        "schema": {
                            "type": "integer",
                            "minimum": 0
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Show all cities",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "$ref": "#/components/schemas/City"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation errors",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ValidationErrors"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Admin Panel"
                ],
                "operationId": "fb1725cfdc2d1032a6a3ed68f0d53e32",
                "parameters": [
                    {
                        "name": "countryId",
                        "in": "query",
                        "description": "Country ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "parentId",
                        "in": "query",
                        "description": "Parent Category ID",
                        "schema": {
                            "type": "integer",
                            "nullable": true
                        }
                    },
                    {
                        "name": "name",
                        "in": "query",
                        "description": "Name of the city",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created city",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "$ref": "#/components/schemas/City"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation errors",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ValidationErrors"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/v1/admin/cities/{id}": {
            "get": {
                "tags": [
                    "Admin Panel"
                ],
                "operationId": "3d1c5bddbcdf09db9ef0de6df58b9a75",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "City ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Show information about the city",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "$ref": "#/components/schemas/City"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation errors",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ValidationErrors"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Admin Panel"
                ],
                "operationId": "083b0432171448f133e66ee7beb0a494",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "City ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "The city has been deleted"
                    },
                    "422": {
                        "description": "Validation errors",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ValidationErrors"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "patch": {
                "tags": [
                    "Admin Panel"
                ],
                "operationId": "e14dd9812dbf4a95b5bd2b26a5a7cf3b",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "City ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "countryId",
                        "in": "query",
                        "description": "Country ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "parentId",
                        "in": "query",
                        "description": "Parent Category ID",
                        "schema": {
                            "type": "integer",
                            "nullable": true
                        }
                    },
                    {
                        "name": "name",
                        "in": "query",
                        "description": "Name of the city",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Show information about updated city",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "$ref": "#/components/schemas/City"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation errors",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ValidationErrors"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "City": {
                "description": "Show information about city",
                "properties": {
                    "id": {
                        "type": "integer"
                    },
                    "country": {
                        "$ref": "#/components/schemas/CountryIdNameAndSlug"
                    },
                    "parentId": {
                        "description": "Integer or null",
                        "type": "integer"
                    },
                    "name": {
                        "type": "string"
                    },
                    "slug": {
                        "type": "string"
                    },
                    "createdAt": {
                        "type": "string"
                    },
                    "updatedAt": {
                        "type": "string"
                    }
                },
                "type": "object",
                "example": [
                    {
                        "id": 1,
                        "country": {
                            "id": 1,
                            "name": "Україна"
                        },
                        "parentId": null,
                        "name": "Київ",
                        "slug": "ukrayina-kiyiv",
                        "createdAt": "2023-12-02 14:51:18",
                        "updatedAt": "2023-12-02 14:56:21"
                    }
                ]
            },
            "CountryIdNameAndSlug": {
                "description": "Show information about country id and name",
                "properties": {
                    "id": {
                        "description": "Transform the resource into an array.",
                        "type": "integer"
                    },
                    "name": {
                        "type": "string"
                    }
                },
                "type": "object"
            },
...
            "ValidationErrors": {
                "description": "Validation Errors",
                "properties": {
                    "message": {
                        "type": "string"
                    },
                    "errors": {
                        "description": "each key describes error message",
                        "properties": {
                            "query or path name": {
                                "type": "array",
                                "items": {}
                            }
                        },
                        "type": "object"
                    }
                },
                "type": "object",
                "example": {
                    "message": "The email field must be a valid email address. (and 3 more errors)",
                    "errors": {
                        "email": [
                            "The email field must be a valid email address."
                        ],
                        "password": [
                            "The password field confirmation does not match.",
                            "The password field must contain at least one uppercase and one lowercase letter.",
                            "The password field must contain at least one number."
                        ]
                    }
                }
            },
...
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "name": "bearerAuth",
                "in": "header",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    }
}